Update AppVeyor config to fix recent errors
* A recent AppVeyor update caused the existing AppVeyor config to fail
when checking make.paws.  It was attempting to install the branch
paws.common (specified in the build_script for only make.paws) before
the general build_script, where it would fail because devtools was not
installed.  Now this step is specified in the after_build step, so it
will only run after the build_script.