Ensure that data from S3 response headers get included in output
* Services like S3 send responses that can contain data in both the body and the header. We unmarshal both and put them in the output object, however currently the data from the header does not actually get included in the output object.

In this update, when unmarshalling data from the body, we instead start with the provided `interface` object, which may already contain data from the header, rather than an empty list. Unmarshalling from the header takes place before the body, and we pass this to the function that unmarshals the body. This ensures that any data unmarshalled from the header gets included when unmarshalling from the body.

We also check that if a given field does not exist in the HTTP response body and there is data already in the output object, we don't overwrite it with a missing value.