Include S3 data from headers in output (#248)
* Services like S3 send responses that can contain data in both the body and the header. We unmarshal both and put them in the output object, however currently the data from the header does not actually get included in the output object.

In this update, when unmarshalling data from the body, we instead start with the provided `interface` object, which may already contain data from the header, rather than an empty list. Unmarshalling from the header takes place before the body, and we pass this to the function that unmarshals the body. This ensures that any data unmarshalled from the header gets included when unmarshalling from the body.

* Make unmarshal-body functions skip fields from that aren't in the body

* Update REST-JSON tests to ignore extra object attributes