Speed up make_sdk()
A great deal of time in make_sdk() is spent in two classes of operations:

1. url_ok()
2. Converting HTML to text or markdown using pandoc

These operations are pretty slow, not dependent on side effects, and their
implementations are pretty stable at this point (maybe?), so they're good
candidates for persistent caching.

This commit uses the cachem package to manage a ./cache subdirectory for
uses of url_ok() and html_to().

On my MBP15 from 2016, I got the following times for make_sdk():

- Cache disabled:         62m17.732s
- With cache, first run:  43m11.440s
- With cache, second run: 10m0.548s

(The "first run" starts with the cache empty, but there are enough
duplicated html_to() calls that it still gets some cache hits.)

If the implementations of url_ok() or html_to() change, you'll need to
throw out the cached values by simply deleting the ./cache directory;
you can use `make clean` for this.

(I also added default arguments for make_sdk; I didn't see a downside to
doing this, and it made iterative profiling a bit less annoying.)