Remove most stray backslashes in docs (#353)
* Use consistent shQuote type

On non-Windows platforms, quoted() gave different behavior (single
instead of double quotes) which caused unnecessary doc diffs. With
this change, you get the same quoting behavior on every platform.

(I think deparse() might be superior to shQuote() anyway, but I'll
leave it like this for now.)

* Remove (most) stray backslashes from docs

Fixes stray '\' characters that appear everywhere in the docs. This
is due to pandoc's overly-aggressive markdown escaping in `markdown`
output mode. With `commonmark` mode pandoc does this an order of
magnitude less, but there are still some left. Grepping for
'\\\\[^a-zA-Z]' in the ./paws/man directory shows them.

This commit doesn't reflect the actual changes to the autogenerated
docs; those will be in the next commit, so this one can be reviewed
separately from all that noise.

* Rebuild package

* Fix tests, upgrade pandoc version used for CI