Remove extra escapes, fix overly long titles (#378)
* Remove most extra escapes. 

Now the only escaping done is for the following cases:
    * All `{` and `}` in plain text to avoid LaTeX errors.
    * Unmatched `{` and `}` in code fragments to avoid Roxygen errors.
    * Unmatched `"`, `'`, or ` in code fragments to avoid Roxygen errors.

Explicitly remove extra escapes added by Pandoc for `_` and for `\` before `{` and `}`. Explicitly remove extra escapes in generated example code.

* Clean up the HTML-to-Markdown pipeline

This had a lot of extra complexity that is no longer needed since moving to Commonmark. In addition, in `convert`, treat all provided documentation as HTML. Previously it would check for first character `<` and if not found, would treat as plain text. However, some documentation had extra whitespace before the HTML began. Now it trims whitespace, and if the text does not begin with an HTML tag, it wraps it in a `<body>` element.

* Fix long help page titles

This was caused by documentation whose first paragraph doesn't end with a period. Previously it would include the next paragraph; now it does not.

* Run Roxygen only for the `cran` folder packages and install from these folders

Run Roxygen only for the packages in the `cran` folder. Previous it also ran for the `paws` folder, which was not needed and takes extra time. In addition, this PR changes the Makefile `install` recipe to install from the `cran` folder as well.