Fix integration tests (#387)
* Fix the integration test make recipe

Make the integration test recipe run tests on each of the CRAN category
packages in the `cran` folder, rather than the legacy paws package in the
`paws` folder. The tests in the legacy paws package no longer work
because we no longer run the Roxygen2/document process on it.

* Fix duplicated lines in integration test output

Delete the call to `context("svc")` in generated integration test files.
`testthat` no longer needs `context()` and now it only results in a
duplicated line in the output.